feature #1613: Check save_as operations
(cherry picked from commit e2d11a0e1629664d5b89e1a7073f743cca91745a)

+++ src/rm/RequestManagerImage.cc
@@ -347,7 +347,7 @@ void ImageClone::request_execute(
     img_usage.add("DATASTORE", ds_id);
     img_usage.add("SIZE", size);
 
-    if (ds_check && (size > avail))
+    if (ds_check && ((unsigned int) size > avail))
     {
         failure_response(ACTION, "Not enough space in datastore", att);
